refactor: move readdirRecursive to its own file
